+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Statsbeat bug fixes
+([#1100](https://github.com/census-instrumentation/opencensus-python/pull/1100))
+
 ## 1.1.1
 Released 2022-01-19