+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Statsbeat bug fixes
 ([#1100](https://github.com/census-instrumentation/opencensus-python/pull/1100))
+- Statsbeat shorten host in network stats
+([#1100](https://github.com/census-instrumentation/opencensus-python/pull/1100))
 
 ## 1.1.1
 Released 2022-01-19