+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+## 1.1.2
+Released 2022-03-02
+
 - Statsbeat bug fixes, shorten host in network stats
 ([#1100](https://github.com/census-instrumentation/opencensus-python/pull/1100))
 - Support statsbeat in EU regions